Update and improve test coverage.
