+++ tests/files/codetree.jl
@@ -1,0 +1,23 @@
+module Alpha
+
+module Bravo
+
+function f1(x)
+  2x
+end
+
+module Charlie
+
+end
+
+end
+
+module Delta
+
+end
+
+end
+
+module Echo
+
+end
