+++ tests/files/s1.jl
@@ -1,0 +1,1 @@
+function f1(); end
