+++ tests/files/s2.jl
@@ -1,0 +1,9 @@
+module Alpha
+function f1(); end
+end
+function f2(); end
+module Bravo
+module Charlie
+function f3(); end
+end
+end
