+++ tests/files/s3.jl
@@ -1,0 +1,4 @@
+module Geometry
+area_circle(radius) = π * r^2
+end
+f1() = "f1"
