+++ tests/files/s4.jl
@@ -1,0 +1,4 @@
+module Alpha
+include("a1.jl")
+include("a2.jl")
+end
