+++ tests/files/s5.jl
@@ -1,0 +1,5 @@
+module Alpha
+module Bravo
+include("a1.jl")
+end
+end
