+++ tests/tests.jl
@@ -1,0 +1,14 @@
+# running the tests:
+# $ julia/tests.jl
+# or:
+# julia> include("tests/tests.jl")
+
+module JuliaSnailTests
+
+using Test
+
+include("../JuliaSnail.jl")
+
+include("./parser.jl")
+
+end
